# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "# Notice: The URL in `Example Request` is currently wrong. The URL\
    \ to use is: `https://api.sandbox.mobilepay.dk/pos/api/v10/...`\n## PoS  V10 API\n\
    \nThe API is split into 5 categories: \n* **Brands**: Look up brands or stores\
    \ based on brand \n* **PointOfSales**: Management of Point of Sales including\
    \ create, delete and querying. Note that a Point of Sale is not editable. \n*\
    \ **Refunds**: Everything needed in order to do a refund (create, cancel and look\
    \ up) \n* **Stores**: Look up store information \n* **Payments**: Everything needed\
    \ in order to initiate and complete a payment. \n\n#### Payment flows\nPayments\
    \ can be started in two ways: as an initiated payment that is immediatedly ready\
    \ for user approval or as a prepared payment that is not yet ready for user approval\
    \ and must be explicitly marked as ready. Once a payment is ready for user approval,\
    \ it is issued to the user. Once the user accepts the payment request, and the\
    \ payment amount has been reserved, the payment must be captured to complete the\
    \ payment.\n\nThe API thus supports the following two payment flows:\n* Initiate\
    \ payment (POST /payments) --> poll until reserved (GET /payments/{paymentId})\
    \ --> capture (POST /payments/{paymentId}/capture) \n* Prepare payment (POST /payments/prepare)\
    \ --> mark as ready (POST /payments/{paymentId}/ready) --> poll until reserved\
...
basePath: "/pos"
schemes:
...
consumes:
...
produces:
...
security: []
paths:
  /api/v10/payments/{paymentId}:
    get:
      tags:
      - "Payments"
      summary: "Lookup a Payment"
      operationId: "QueryPayment"
      consumes: []
      produces:
      - "application/json"
      parameters:
      - name: "paymentId"
        in: "path"
        description: "PaymentId"
        required: true
        type: "string"
        format: "uuid"
      - name: "Authorization"
        in: "header"
        description: "Integrator's Bearer Token"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-Id"
        in: "header"
        description: "Integrator's MobilePay Client Id (Must be a valid GUID)"
        required: true
        type: "string"
        format: "uuid"
      - name: "X-MobilePay-Client-System-Name"
        in: "header"
        description: "Integrator's Certified System Name"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-System-Version"
        in: "header"
        description: "Integrator's Certified System Version"
        required: true
        type: "string"
      responses:
        "200":
          description: "The Payment"
          examples:
            application/json:
              paymentId: "1cbfff94-3d17-4dc1-b667-5280e1ce50f9"
              posId: "c0000a0f-68b8-4759-847b-08d5284c344c"
              orderId: "ORDER-12345"
              amount: 12.5
              currencyCode: "DKK"
              restrictions:
                debitCardDisallowed: false
                creditCardDisallowed: false
                userMinimumAge: 18
              merchantPaymentLabel: "PaymentLabel"
              plannedCaptureDelay: "None"
              status: "Paired"
              customerToken: "41e519d3b5ac1a228bd15fab8958f1d9b4ee28eb"
              customerReceiptToken: "671c4da4859a4639a5453120d791aac0"
              loyaltyIds:
              - "123456"
              pollDelayInMs: 100
          schema:
            $ref: "#/definitions/QueryPaymentResponse"
        "401":
          description: "Unauthorized"
        "403":
          description: "Forbidden"
        "404":
          description: "Payment not found"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: "Internal server error"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-swagger-router-controller: "Payments"
  /api/v10/payments:
    get:
      tags:
      - "Payments"
      summary: "Lookup Payments for a given filter"
      operationId: "QueryPaymentIds"
      consumes: []
      produces:
      - "application/json"
      parameters:
      - name: "posId"
        in: "query"
        description: "Restricts Payments returned to those with a specific PosId"
        required: false
        type: "string"
        format: "uuid"
      - name: "orderId"
        in: "query"
        description: "Restricts Payments returned to those with a specific orderId"
        required: false
        type: "string"
      - name: "active"
        in: "query"
        description: "Restricts Payments returned to those that are active"
        required: false
        type: "boolean"
        default: false
      - name: "Authorization"
        in: "header"
        description: "Integrator's Bearer Token"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-Id"
        in: "header"
        description: "Integrator's MobilePay Client Id (Must be a valid GUID)"
        required: true
        type: "string"
        format: "uuid"
      - name: "X-MobilePay-Client-System-Name"
        in: "header"
        description: "Integrator's Certified System Name"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-System-Version"
        in: "header"
        description: "Integrator's Certified System Version"
        required: true
        type: "string"
      responses:
        "200":
          description: "List of paymentIds satisfied by the filter"
          examples:
            application/json:
              paymentIds:
              - "1cbfff94-3d17-4dc1-b667-5280e1ce50f9"
          schema:
            $ref: "#/definitions/QueryPaymentIdsResponse"
        "204":
          description: "No paymentIds satisfied by the filter exist"
        "401":
          description: "Unauthorized"
        "403":
          description: "Forbidden"
        "409":
          description: "Filter not specific enough"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: "Internal server error"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-swagger-router-controller: "Payments"
    post:
      tags:
      - "Payments"
      summary: "Initiate a new Payment"
      operationId: "InitiateReservationPayment"
      consumes:
      - "application/json-patch+json"
      - "application/json"
      - "text/json"
      - "application/*+json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "request"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/InitiateReservationPaymentRequest"
      - name: "Authorization"
        in: "header"
        description: "Integrator's Bearer Token"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-Id"
        in: "header"
        description: "Integrator's MobilePay Client Id (Must be a valid GUID)"
        required: true
        type: "string"
        format: "uuid"
      - name: "X-MobilePay-Client-System-Name"
        in: "header"
        description: "Integrator's Certified System Name"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-System-Version"
        in: "header"
        description: "Integrator's Certified System Version"
        required: true
        type: "string"
      - name: "X-MobilePay-Idempotency-Key"
        in: "header"
        description: "Used to ensure retried calls are handled correctly"
        required: true
        type: "string"
      responses:
        "200":
          description: "Payment initiated"
          examples:
            application/json:
              paymentId: "1cbfff94-3d17-4dc1-b667-5280e1ce50f9"
          schema:
            $ref: "#/definitions/PaymentInitiatedResponse"
        "400":
          description: "Bad Request. See error response for details"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "401":
          description: "Unauthorized"
        "403":
          description: "Forbidden"
        "409":
          description: "Payment could not be initiated. See error response for details"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: "Internal server error"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-swagger-router-controller: "Payments"
  /api/v10/payments/prepare:
    post:
      tags:
      - "Payments"
      summary: "Prepare a new Payment that is not yet ready for User approval"
      operationId: "PrepareReservationPayment"
      consumes:
      - "application/json-patch+json"
      - "application/json"
      - "text/json"
      - "application/*+json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "request"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/PrepareReservationPaymentRequest"
      - name: "Authorization"
        in: "header"
        description: "Integrator's Bearer Token"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-Id"
        in: "header"
        description: "Integrator's MobilePay Client Id (Must be a valid GUID)"
        required: true
        type: "string"
        format: "uuid"
      - name: "X-MobilePay-Client-System-Name"
        in: "header"
        description: "Integrator's Certified System Name"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-System-Version"
        in: "header"
        description: "Integrator's Certified System Version"
        required: true
        type: "string"
      - name: "X-MobilePay-Idempotency-Key"
        in: "header"
        description: "Used to ensure retried calls are handled correctly"
        required: true
        type: "string"
      responses:
        "200":
          description: "Payment prepared"
          examples:
            application/json:
              paymentId: "1cbfff94-3d17-4dc1-b667-5280e1ce50f9"
          schema:
            $ref: "#/definitions/PaymentInitiatedResponse"
        "400":
          description: "Bad Request. See error response for details"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "401":
          description: "Unauthorized"
        "403":
          description: "Forbidden"
        "409":
          description: "Payment could not be prepared. See error response for details"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: "Internal server error"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-swagger-router-controller: "Payments"
...
  /api/v10/payments/{paymentId}/capture:
    post:
      tags:
      - "Payments"
      summary: "Capture a Payment. Only reserved payments can be captured."
      operationId: "CapturePayment"
      consumes:
      - "application/json-patch+json"
      - "application/json"
      - "text/json"
      - "application/*+json"
      produces:
      - "application/json"
      parameters:
      - name: "paymentId"
        in: "path"
        description: "PaymentId"
        required: true
        type: "string"
        format: "uuid"
      - in: "body"
        name: "request"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/PaymentCaptureRequest"
      - name: "Authorization"
        in: "header"
        description: "Integrator's Bearer Token"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-Id"
        in: "header"
        description: "Integrator's MobilePay Client Id (Must be a valid GUID)"
        required: true
        type: "string"
        format: "uuid"
      - name: "X-MobilePay-Client-System-Name"
        in: "header"
        description: "Integrator's Certified System Name"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-System-Version"
        in: "header"
        description: "Integrator's Certified System Version"
        required: true
        type: "string"
      responses:
        "200":
          description: "Payment has been captured"
        "400":
          description: "Bad Request. See error response for details"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "401":
          description: "Unauthorized"
        "403":
          description: "Forbidden"
        "404":
          description: "Payment not found"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "409":
          description: "Payment could not be captured. See error response for details"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: "Internal server error"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-swagger-router-controller: "Payments"
  /api/v10/payments/{paymentId}/cancel:
    post:
      tags:
      - "Payments"
      summary: "Cancel a payment. A payment cannot be cancelled once it has been captured."
      operationId: "CancelPayment"
      consumes: []
      produces:
      - "application/json"
      parameters:
      - name: "paymentId"
        in: "path"
        description: "PaymentId"
        required: true
        type: "string"
        format: "uuid"
      - name: "Authorization"
        in: "header"
        description: "Integrator's Bearer Token"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-Id"
        in: "header"
        description: "Integrator's MobilePay Client Id (Must be a valid GUID)"
        required: true
        type: "string"
        format: "uuid"
      - name: "X-MobilePay-Client-System-Name"
        in: "header"
        description: "Integrator's Certified System Name"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-System-Version"
        in: "header"
        description: "Integrator's Certified System Version"
        required: true
        type: "string"
      responses:
        "200":
          description: "Payment cancelled"
        "401":
          description: "Unauthorized"
        "403":
          description: "Forbidden"
        "404":
          description: "Payment not found"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "409":
          description: "The payment could not be cancelled. See error response for\
            \ details"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: "Internal server error"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-swagger-router-controller: "Payments"
...
definitions:
  QueryPaymentResponse:
    type: "object"
    required:
    - "orderId"
    - "paymentId"
    - "pollDelayInMs"
    - "posId"
    - "status"
    properties:
      paymentId:
        type: "string"
        format: "uuid"
      posId:
        type: "string"
        format: "uuid"
      orderId:
        type: "string"
      amount:
        type: "number"
        format: "double"
      currencyCode:
        type: "string"
        enum:
        - "DKK"
        - "EUR"
      restrictions:
        $ref: "#/definitions/PaymentRestrictions"
      merchantPaymentLabel:
        type: "string"
      plannedCaptureDelay:
        type: "string"
        enum:
        - "None"
        - "LessThan24Hours"
        - "LessThan14Days"
      status:
        type: "string"
        enum:
        - "Prepared"
        - "Initiated"
        - "Paired"
        - "IssuedToUser"
        - "Reserved"
        - "CancelledByUser"
        - "CancelledByClient"
        - "CancelledByMobilePay"
        - "Captured"
      customerToken:
        type: "string"
      customerReceiptToken:
        type: "string"
      loyaltyIds:
        type: "array"
        description: "If LoyaltyIds is null, MobilePay has not yet looked up loyalty\
          \ for this user. Retry the call. \\\r\nIf LoyaltyIds is an empty list, this\
          \ user is not enrolled in the Merchants loyalty programme."
        items:
          type: "string"
      pollDelayInMs:
        type: "integer"
        format: "int32"
  PaymentRestrictions:
    type: "object"
    properties:
      debitCardDisallowed:
        type: "boolean"
      creditCardDisallowed:
        type: "boolean"
      userMinimumAge:
        type: "integer"
        format: "int32"
    description: "Allow setting restriction on the payment methods and user for a\
      \ payment"
  ErrorResponse:
    type: "object"
    properties:
      code:
        type: "string"
        readOnly: true
      message:
        type: "string"
        readOnly: true
      correlationId:
        type: "string"
        readOnly: true
  QueryPaymentIdsResponse:
    type: "object"
    required:
    - "paymentIds"
    properties:
      paymentIds:
        type: "array"
        items:
          type: "string"
          format: "uuid"
  InitiateReservationPaymentRequest:
    type: "object"
    required:
    - "amount"
    - "currencyCode"
    - "orderId"
    - "plannedCaptureDelay"
    - "posId"
    properties:
      posId:
        type: "string"
        format: "uuid"
      orderId:
        type: "string"
      amount:
        type: "number"
        format: "double"
      currencyCode:
        type: "string"
        enum:
        - "DKK"
        - "EUR"
      restrictions:
        $ref: "#/definitions/PaymentRestrictions"
      merchantPaymentLabel:
        type: "string"
      plannedCaptureDelay:
        type: "string"
        enum:
        - "None"
        - "LessThan24Hours"
        - "LessThan14Days"
    example:
      posId: "c0000a0f-68b8-4759-847b-08d5284c344c"
      orderId: "ORDER-12345"
      amount: 12.5
      currencyCode: "DKK"
      restrictions:
        debitCardDisallowed: false
        creditCardDisallowed: false
        userMinimumAge: 18
      merchantPaymentLabel: "PaymentLabel"
      plannedCaptureDelay: "None"
  PaymentInitiatedResponse:
    type: "object"
    required:
    - "paymentId"
    properties:
      paymentId:
        type: "string"
        format: "uuid"
  PrepareReservationPaymentRequest:
    type: "object"
    required:
    - "orderId"
    - "posId"
    properties:
      posId:
        type: "string"
        format: "uuid"
      orderId:
        type: "string"
    example:
      posId: "c0000a0f-68b8-4759-847b-08d5284c344c"
      orderId: "ORDER-12345"
...
  PaymentCaptureRequest:
    type: "object"
    properties:
      amount:
        type: "number"
        format: "double"
    example:
      amount: 12.5
...